+++ src/components/FrameApi.vue
@@ -27,7 +27,7 @@ export default {
         this.response = new Promise(resolve => resolve(response));
         this.$emit("success", { data: this.response });
       } catch (error) {
-        // We probably came from `api.service.js` and need to logout.
+        // We probably came from `api.js` and need to logout.
         // The error is undefined and we quit here.
         if (error === undefined) return;
