+++ src/services/clock.js
@@ -1,4 +1,4 @@
-import ApiService from "@/services/api.service";
+import ApiService from "@/services/api";
 import store from "@/store";
 
 function mapApiResponse(response) {
