+++ src/services/shift.js
@@ -1,7 +1,7 @@
-import ApiService from "@/services/api.service";
+import ApiService from "@/services/api";
 import store from "@/store";
 
-// class ContractError extends Error {
+// class ShiftError extends Error {
 //   constructor(errorCode, message) {
 //     super(message);
 //     this.name = this.constructor.name;
@@ -14,15 +14,19 @@ function mapApiResponse(response) {
   return {
     uuid: response.id,
     user: response.user,
-    name: response.name,
-    date: { start: response.start_date, end: response.end_date },
-    hours: response.hours
+    contract: response.contract,
+    date: { start: response.started, end: response.stopped },
+    type: response.type,
+    note: response.note,
+    tags: response.tags,
+    exported: response.was_exported,
+    reviewed: response.was_reviewed
   };
 }
 
-const BASE_URL = "/api/contracts/";
+const BASE_URL = "/api/shifts/";
 
-const ContractService = {
+const ShiftService = {
   create: async function(data) {
     const requestData = {
       method: "post",
@@ -33,10 +37,10 @@ const ContractService = {
     return new Promise((resolve, reject) => {
       return ApiService.customRequest(requestData)
         .then(response => {
-          const contract = mapApiResponse(response.data);
-          store.dispatch("contract/addContract", contract);
+          const shift = mapApiResponse(response.data);
+          store.dispatch("shift/addShift", shift);
 
-          return resolve(contract);
+          return resolve(shift);
         })
         .catch(error => {
           reject(error);
@@ -47,8 +51,9 @@ const ContractService = {
     return new Promise((resolve, reject) => {
       return ApiService.get(BASE_URL + `${uuid}`)
         .then(response => {
-          const contract = mapApiResponse(response.data);
-          const newResponse = { ...response, contract };
+          const shift = mapApiResponse(response.data);
+          // TODO: we might need to use `{ ...response, data: shift }` here!?
+          const newResponse = { ...response, shift };
 
           return resolve(newResponse);
         })
@@ -74,10 +79,10 @@ const ContractService = {
     return new Promise((resolve, reject) => {
       return ApiService.patch(`${BASE_URL}${uuid}/`, data)
         .then(response => {
-          const contract = mapApiResponse(response.data);
-          store.dispatch("contract/updateContract", contract);
+          const shift = mapApiResponse(response.data);
+          store.dispatch("shift/updateShift", shift);
 
-          return resolve(contract);
+          return resolve(shift);
         })
         .catch(error => {
           reject(error);
@@ -97,4 +102,4 @@ const ContractService = {
   }
 };
 
-export default ContractService;
+export default ShiftService;
