+++ src/views/ViewShiftForm.vue
@@ -21,7 +21,7 @@
 <script>
 import ShiftForm from "@/components/shifts/ShiftForm";
 import { Shift } from "@/models/ShiftModel";
-import ShiftService from "@/services/shift.service";
+import ShiftService from "@/services/shift";
 import FrameApi from "@/components/FrameApi";
 
 export default {
